Backend/RAG/requirements.txt
langchain==0.2.5
langchain-google-genai==0.0.6
google-generativeai==0.3.2
langchain-community==0.2.5
faiss-gpu==1.7.2
python-dotenv==1.0.1
fastapi
pillow
